Refactor tasks into duke.task package
