+++ src/main/java/duke/task/Todo.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.task;
 
 public class Todo extends Task {
     public Todo(String description) {
